requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.0
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.14.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==7.1.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
keras==3.10.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pillow==11.3.0
protobuf==6.33.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
rich==14.2.0
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tifffile==2024.8.30
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.4
wrapt==2.0.1
zipp==3.19.2
